requirements.txt
automata-lib
visual-automata
pandas
graphviz